pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "energy-mix"
version = "0.1.4"
description = "Twitter bot tweeting energy mix of electricity generated in Czechia in past hour."
readme = "README.md"
license = {file = "LICENSE"}

dependencies = [
  "python-dotenv~=1.0.1",
  "requests~=2.32.3",
  "tweepy~=4.14.0",
  "xmltodict~=0.13.0",
]

classifiers = [
  "Programming Language :: Python"
]

[project.optional-dependencies]
dev = [
    "pytest~=8.2.2",
    "ruff~=0.5.2",
]

[project.scripts]
energy-mix = "energy_mix.__main__:main"
